make file support for running in docker
