+++ requirements/requirements_api.txt
@@ -3,6 +3,7 @@ aiohttp-swagger3==0.7.4
 astropy==5.2.1
 bcrypt==4.0.1
 cryptography==39.0.1
+deepdiff==6.2.3
 gunicorn==20.1.0
 jinja2>=2.11.2
 matplotlib==3.7.0
