+++ requirements/requirements_ingester_macos.txt
@@ -5,6 +5,7 @@ beautifulsoup4==4.11.2
 confluent-kafka==2.0.2
 cryptography==39.0.1
 dask==2023.2.0
+deepdiff==6.2.3
 distributed==2023.2.0
 fastavro==1.6.0 #breaking changes in newer versions
 fire==0.5.0
